Run ReflectionAnalysis in ModelSpecGenerator
Summary: The current way of detecting reflections in Class Merging is still some what adhoc. We look at all const-class references in MergeabilityCheck, but it's not complete. In this change, we use ReflectionAnalysis to detect additional reflections against merging targets in ModelSpecGenerator. In this way, we should be able to detect reflective calls like `Object.getClass()` and `Class.forName()` against anonymous classes.

Reviewed By: NTillmann

Differential Revision: D40405421

fbshipit-source-id: aa4ada7ca0c0e4f8504b7383942e09c784d44c00