Improve enum ordinal analysis precision.
Summary:
Currently it looks at all static fields of the enum's type. But this is too broad, as it's not uncommon to define an alias of an enum as a static field. We end up rejecting some enums.

The ENUM access flag can be used to track which fields are actually the enum variants, and only check those.

Reviewed By: agampe

Differential Revision: D40419002

fbshipit-source-id: e976775c0c9d71a3da8e816a5bb94358869bc174