Check all branches in optimize enums tests.
Summary:
The optimize enums pass is sensitive to syntactic structure. Unfortunately the tests replicate the same syntactic sensitivity, making improving the pass difficult.

This diff uses match flow to extract _all_ branches that compare against a const, not just switches or (specifically-structured) if-trees. While we're here, it also tracks if the source of the comparison comes from an aget or invoke virtual - this lets the test assert that the constant is coming an expected place.

A side effect of this is that we no longer (accidentally?) merge the default branch target with a switchmap/ordinal compare that shares the same target. This is why some set comparisons have an additional expected ordinal, covering the full ordinal range.

Reviewed By: thezhangwei

Differential Revision: D40419005

fbshipit-source-id: 09cb0bf88de5e35287c5100cad35f0b5a0104305