Generalize optimize enums switchmap removal.
Summary:
The existing implementation is structure sensitive, finding only switches and what it deems switch-equivalent if-trees. Unfortunately the latter is easily defeated, and even when it does match the if-tree is turned back into a switch.

This uses match flow to perform the removal. The two immediate benefits are that it handles more places where a switchmap is used, and no longer modifies the underlying branching structure (i.e., if-statements are left as if-statements).

The simplest way to think of the transform (for me at least) is that we are turning `f(x) == C` into `x == f_inv(C)`, where f is a bijection and C is a known constant. What the comparisons are for and where they go is irrelevant to us, hence the simplicity of the approach. We just want to know where some `x` is flowing into an `f()` and being compared against a constant `C`.

This is not a general purpose pass for the above idea, it is limited to switchmaps as the mapping function. But framing it this way removes most of the existing complexity.

Reviewed By: thezhangwei

Differential Revision: D40419000

fbshipit-source-id: 918ff788f20bd14df676d49e43e8feaec02065e5