Optimize Kotlin enums.
Summary:
In vanilla Java, switching over a null enum gives a NPE on the `ordinal()` call. In Kotlin, a nullable enum can be given its own case.

This is implemented by using -1 as the "switch map" result for null. To handle this we must accept more than one path reaching the conditional with either an aget result or a const -1 as an argument.

Reviewed By: thezhangwei

Differential Revision: D40419001

fbshipit-source-id: 1bcb4ec7f00a8299eea8ed438e4912da206d3534