include StringAnalyzer, in particular for call-site summaries
Summary:
All StringDomain values arise from `const-string` (and static fields initialized with a constant string), which according to the JVM must be internalized:
```
The Java programming language requires that identical string literals (that is, literals that contain the same sequence of code points) must refer to the same instance of class String (JLS §3.10.5).
```

With this guarantee about their identity, we can leverage them fully.

Reviewed By: beicy

Differential Revision: D37977787

fbshipit-source-id: 290f423e88178a0508e4474e7b4cb05deee30b2d