exclude joined array types
Summary:
Unfortunately, Java arrays lack a common (non-object) base type, so joining array types doesn't preserve the array-ness of types. And this makes the verification of array-length unhappy.

In this diff, we conservatively don't dedup blocks if we might have joined array types, but couldn't identify a particular joined array type, and the value is used in an array instruction.

Reviewed By: wsanville

Differential Revision: D40814258

fbshipit-source-id: 30b211d17baef5a4ddcb4efe7df1b6832f2e4d99