Do the right thing for invoke-super
Summary:
According to the spec, when resolving an invoke-super, the search much start from the calling method's class. Relying on the declaring type of the method mentioned in the invoke-super instruction is brittle. This is getting fixed here.

(Eventually, resolve_method should have an assertion that requires a caller when doing an super-resolution.)

Reviewed By: beicy

Differential Revision: D41013685

fbshipit-source-id: 2ff9883d4f8d4672c008b230370c45e0d78a2573