Properly identify virtual fallback when merging a complex virtual scope
Summary:
When merging virtual methods within the same virtual scope, we could have a base implementation higher in the inheritance hierarchy that's partially overridden. That is some mergeables do not override the base implementation. In that case we need to fallback to the base one, and emit a `invoke-super` in the merged virtual dispatch switch.

In this scenario, the most common case is that the top_def of the virtual scope is a concrete method and is the correct virtual fallback. However, that's not always the case. In an interface virtual scope, we can have a non-def miranda at the top of the virtual scope. We need to keep scanning the virtual scope to find the correct virtual fallback. This change was on our todo list. Now it's uncovered by an OSS issue.

Please see the attached tests for a detailed illustration of the problem being addressed here.

Reviewed By: ssj933

Differential Revision: D41176242

fbshipit-source-id: 454228e3db892f73cebee1002d5ecb0d6d5fd0d1