Validate class specs in betamaps properly
Summary:
Check that all betamap entries actually end in `.class`.

(This would have prevented a mistake in prod inclusion automation, where the `Lpath/to/classname;` format was used instead of the `.class` suffix.)

Also removed the `exit` call and assertion, replacing both with a `RedexException` throw. A raw `exit` is difficult to test, and assertions should be reserved for internal errors (i.e. Redex bugs), not errors in user input.

And building on that, add a test that an invalid class spec in a betamap will be rejected.

Reviewed By: wsanville

Differential Revision: D41230639

fbshipit-source-id: 5f4d6fd54153271070b286a5907fdc27c83a0896