Allow to generate shorter names for outlined methods (#743)
Summary:
This is a relatively small change and does not change the default behavior.

``InstructionSequenceOutlinerPass`` generates names with stable hashes for outlined methods, this PR adds an option in config to allow whether to generate obfuscated names using ObfuscateUtil API. By default, this should be taken care of when we run ``ObfuscatePass``, however, this change could be useful if some of us cannot enable ``ObfuscatePass`` due to various reasons.

Looking forward to the feedback on this!

Bests,
Umar

Pull Request resolved: https://github.com/facebook/redex/pull/743

Reviewed By: agampe

Differential Revision: D41387047

Pulled By: NTillmann

fbshipit-source-id: 8ac3d8d306bf9efe7e818b207d845a5e4438329a