Emit invocations of Unchecked methods
Summary:
Use checked and `Unchecked` methods for transferring coverage bits at method exit. The general structure is
```
if (onMethodExit(0, ...)) {
  onMethodExitUnchecked(1, ...);
  ...
}
```
This ensures that a flipping value of the enabled state does not lead to half the state being stored.

NOTE: To be fully consistent, `onMethodBegin` should actually be the defining state. However, for now the local on-exit block is hopefully enough.

Update tests.

NOTE: The hit count code will be updated later.

Reviewed By: jimmycFB

Differential Revision: D41454190

fbshipit-source-id: d84d285aa382b6018a0557d9e6ecfc9db2076ec3