Add source block idom maps to redex-instrument-metadata.zip
Summary:
Create a function to serialize immediate dominator mappings in dot format separated by ';'.  E.g. `1->0;2->1;3->2`.

Instead of naively writing this out for every method with source blocks, eliminate some redundancy as follows:

* Write unique serialized idom mappings to `unique-idom-mappings.txt`.

* For all methods with source blocks, write the line index of the idom mapping in `unique_idom_mappings.txt` to `redex-source-block-idom-maps.csv`.
* The order of entries in `redex-source-block-idom-maps.csv` corresponds with the order in `redex-source-blocks.csv`.

This increases the size of uncompressed instrumentation metadata by ~46MB for FB4A, which results in a ~5MB increase in compressed size (~80MB -> ~85MB).  This seems acceptable to me.

The purpose of this is to enable `ddata` to perform early flow violation checks.

Reviewed By: agampe

Differential Revision: D41417471

fbshipit-source-id: c29909b13498aea189ffd857004f9dd7925e9410