Expand intra-dex inlininig capabilities
Summary: Based on NTillmann's diff and idea (D40724161). Currently, the pass will not inline a (caller, callee) pair if the methods reside in different dexes. This is overly conservative, as we only need to make sure that we don’t introduce new references in the caller’s dexes. Next diff makes this more conservative by forbidding callees with init-class instructions.

Reviewed By: NTillmann

Differential Revision: D41401979

fbshipit-source-id: f2c82308732766fd00fa5a14e0d2621e75bd43b2