legacy_perf_logic option
Summary:
The new option is on by default, which represents the existing behavior. Thus, this is a behavior preserving change.

The current behavior is weird, to say the least:
- only methods in "perf-sensitive" classes are considered as perf-sensitive, which is essentially only classes that have been ordered by the InterDex pass
- of those methods, the legacy logic only treats those as perf sensitive which are *not popular*. This is most likely the exact inverse of what's beneficial.

When the option will be turned off, the new behavior would be...
- if method-profiles are present, treat all popular methods as perf-sensitive.
- Otherwise, treat all methods of perf-sensitive classes as perf-sensitive.

(Strings accessed via const-string instructions in perf-sensitive methods are never deduplicated.)

Reviewed By: agampe

Differential Revision: D41668626

fbshipit-source-id: b7798660cae3536ac8a9079f07680bbdd5350314