Ensure proper flags for type strings
Summary:
Conceptually, the code as written before this diff is wrong; if new strings are added to the pool they are put at the end without regards for sorting. If the input data really was sorted and had the flag set, we'd ruin that invariant.

In practice I don't see any type string pools coming in as sorted, so the wrong behavior is probably not harming anything. Still, we should fix the oversight.

Reviewed By: NTillmann

Differential Revision: D42481935

fbshipit-source-id: 7da911799712685543314666b49df40ce2773356