Turn off attr dedup in test cases
Summary: The test case for this was quite useful to make sure our logic was sensible in the face of hypothetical inputs and config options that we are not using today, but, it turns out that we have some tricky issues to work through to get this to work properly on app bundles. Turn off this type in test cases so we do not block any changes like toolchain updates.

Reviewed By: agampe

Differential Revision: D42551042

fbshipit-source-id: fb36cc79be1c4e4754a9ec09fc83a05a38870d65