Avoid use of deprecated std::iterator
Summary:
This is a follow-up to D42637996.

`std::iterator` usage is deprecated in C++, and recent minimal version for toolchain update (D42434164) boosted severity of this warning, turning it into error and breaking the build.

This diff does an equivalent code change, which is compliant with the modern C++, thus unbreaking the build.

Reviewed By: NickGerleman

Differential Revision: D42638863

fbshipit-source-id: 16f1d38bdd33986a6e0a445863d03ce3b67fbfc3