Fix visibility issues
Summary:
`ObfuscatePass` "sharpens" calls to their definition, seemingly in an effort to not have to create new method refs (that would replace existing ones, so should be fine with the ref limit). This can lead to visibility problems, when the base class is not visible to the caller.

Make classes public when necessary.

Add a test.

Reviewed By: thezhangwei

Differential Revision: D43012633

fbshipit-source-id: 47d0c20795625f4e2ab392271661e535a66ebf1f