Don't do concurrent destruction in global scope
Summary:
The order of global scope destruction is hard to control.
Fast concurrent destruction of concurrent collection has a depedencies on threading and the sparta-workqueue.
To avoid issues with the destruction order between all those entities, this diff makes it so that by default concurrent destruction is disabled.
The new `ConcurrentContainerConcurrentDestructionScope` can be used at the top-level function of the application to allow for controlled fast concurrent destruction.

By far most concurrent containers fall within that top-level function scope, only very few are at the global scope.

Reviewed By: wsanville

Differential Revision: D43064251

fbshipit-source-id: 2650a1a1f0596874c8b97964f0cdcda7937c2b2c