Surrender to Android limitations
Summary:
Did you know that this perfectly legal Java program is eventually rejected by the Android verifier (at least on Android 12)??
```
public class Program {
  static interface I {}
  static class C implements I {}
  static class D implements I {}
  static void foo(I[] a){}
  public static void main(String[] args) {
    I[] a;
    if (args.length == 0) a = new C[0];
    else if (args.length == 1) a = new D[0];
    else return;
    foo(a);
  }
}
```

This is because the Android verifier can not properly join arrays of reference types in the presence of interfaces, but assumes that it can when validating invocation arguments.

A recent improvement to the DedupBlocks pass may synthesize such code. This is being prevented by this diff.

Longer-term, we should teach our own verifier about this quirk, and audit other passes to make sure they also respect this quirk.

Reviewed By: beicy

Differential Revision: D43208323

fbshipit-source-id: aa8f776bfe8b5cff370c52c53c7c9a7dec89a40a