Fix forwarding
Summary:
This fixes store-load forwarding in the presence of throw-edges.
In a nutshell, the bug was that forwarding was installed when visiting the store instruction (sput/iput/aput, and new-array which "stores" length), not taking into account whether execution might continue along a throw-edge, in which case the value won't actually be stored (and a move-instruction to a fresh register that we'd insert right *after* a *put instruction won't assign to that register when the store instruction throws).

This is getting fixed here by only installing the forwarding when analyzing the no-throw continuation of a store instruction.

Reviewed By: agampe

Differential Revision: D43523515

fbshipit-source-id: c1eba63079eb96acbbe5e466d909ec4a89154991