More flexibility when rebuilding string pools
Summary:
This change makes the `rebuild_string_pool` function behave a bit more sensibly based on the parameters.

Before, the `kept_old_to_new` map was just consulted as a set, and if a string was kept it would be emitted sequentially, even if the "new" value said otherwise. With this change, an arbitrary ordering can be accomplished (if we want that).

Reviewed By: NTillmann

Differential Revision: D42483073

fbshipit-source-id: 4f6562ff2553e88a06bcbdc33b73805802198858