Make deterministic
Summary:
We have recently observed non-determinism in IPCP.
In my investigation I have seen that non-determinism in the call graph (order of edges) causes non-determinism in the final result of IPCP.
This diff effectively addresses this problem by making the call graph deterministic.
(To make up for the time lost on extra sorting, I made the locking more fine-grained.)

However, even after spending a significant amount on time on this whole problem, it is not clear to me how non-determinism in the order of the edges actually affects the outcome of IPCP. It shouldn't matter, but it does. So I fear that there's another problem buried somewhere. But that's for another time.

Reviewed By: agampe

Differential Revision: D43968790

fbshipit-source-id: 4f2e17f0f30e2467a9d7704b1a291cf1dca19085