Identify and separate "blanket native" keep rules
Summary:
A "blanket native" keep rule is a Proguard keep rule which keeps all native methods and/or their parent classes.  For example:

```
   -keepclasseswithmembers class * {
      native <methods>;
    }
```

These sorts of rules are sprinkled throughout many proguard config files used by apps and components.  Their intent is to prevent breaking JNI registration.  However, they make reachability analysis of native methods difficult.

Here, we identify them and move them to a logically separate list of keep rules so that (in a dependent diff) we can determine their effects on reachability in isolation.

Physically, we move their pointers to the end of the `KeepSpecSet`'s ordered vector and store the iterator pointing at the beginning in ProguardConfiguration.  A new `stable_partition` API is added to `KeepSpecSet` to achieve this.

`ProguardMatcher::process_keep` has been changed to take a pair of iterators, and the "blanket native" rules are processed after .

Also, noted that the construction of `ProguardConfiguration` is not well-encapsulated and left a comment + task.  Added `const` to post-construction references downstream, as it shouldn't be modified later.

Reviewed By: agampe

Differential Revision: D44072288

fbshipit-source-id: d4cfc491c038fb47d38c47870d6b3b7686b089b6