Split true virtual rtype specialization validation from the actual update
Summary:
Instead of analyzing/validating various cases for true virtual, and updating them in the same pass, we now store all the validated virtual scope roots in a `virtual_roots` map.

The size of the `virtual_roots` map is actual quite small. The restructure helps to parallelize the first step that processes the large number of candidates.

Reviewed By: NTillmann

Differential Revision: D44154028

fbshipit-source-id: 23a977833d730588548106f67cfa45aeba829019