Rename MethodSimilarityOrderer to MethodSimilarityGreedyOrderer
Summary: We are introducing a new similarity re-ordering scheme later in the stack, so let's rename the existing one to differentiate between the two. While both are re-ordering by similarity, this approach is computing scores between code items and greedily chooses the next candidate, while the new approach is clustering code items that are likely to compress better together.

Reviewed By: NTillmann

Differential Revision: D44298046

fbshipit-source-id: ddccdc483e1b650ac33875bcc099decb13c3ada2