Improve method proto collision check for bridge methods
Summary:
# Context

The method proto update collision check in `SingleImplPass` misses a case for bridge methods.
Here is an example:
```
interface Intf { Intf setup(Intf i); }

class Impl {
  Impl setup(Intf i) {
     return this;
  }
}
```
The bridge method on class `Impl` has the signature of `Intf setup(Intf)`.
The actual implementation method on class `Impl` is `Impl setup(Intf)`.
The collision only exists if we update the proto for both methods.

# Mitigation
For the bridge method, we can perform additional collision check by only update the rtype on the proto (`Impl setup(Intf)` in the above example) just to be conservative.

Reviewed By: beicy

Differential Revision: D44564042

fbshipit-source-id: d801e3e49ef960f1beac0914f06885de2c2ff21d