Get rid of std::bind from SpartaWorkQueue
Summary:
It's unnecessary, as `std::thread` has a variadic constructor, to which all args can be passed.  `std::bind` requires an allocation upfront and is slower to invoke.

We saw `std::thread` construction at this callsite show up as very hot in a small app redex trace recently.  Could not repro it locally, but did notice this unnecessary use of `std::bind` and it seemed worth fixing.

Reviewed By: agampe

Differential Revision: D44883711

fbshipit-source-id: ccc3519238f6341cdec67adbac56c843e0bbd118