Introducing "properties"
Summary:
These properties reflect the state of the code, and possibly some other auxiliary state.
Every pass "interacts" with these properties by optionally...
- requiring a property
- preserving a property
- establishing a property

When a pass neither establishes nor preserves a property, then it effectively "destroys" the property.

There are two layers of (optional) checking:
- quick static checking before passes run to validate that, given an initial property state, and going through all the interactions of the configured active passes, we eventually get to the required final state.
- dynamic deep checking of properties, ensuring that each pass indeed preserved or established properties as promised

To start with, this diff defines the NoInitClassInstruction property, annotates all the relevant passes that either establish or preserve this property, and implements the corresponding (deep) checker.

Reviewed By: jimmycFB

Differential Revision: D44822810

fbshipit-source-id: 030fdb788806eb7d24b37176f93819587c855541