A property for `MakePublicPass`
Summary:
Introduce "negative" properties. These properties are defined by returning `true` for `is_negative`. The effect is two-fold:
* Negative properties are by default preserved.
* Negative properties must not be in the final state.

When using these, an established property will not automatically be destroyed and the checker will complain if the property is not removed  before the final state.

Introduces `NeedsEverythingPublic` property that models the current requirement of some passes to have some other pass (`MakePublicPass`) clean up at the end, fixing introduced access violations.

Annotations of passes that introduce the property is a follow-up item.

Reviewed By: NTillmann

Differential Revision: D44925308

fbshipit-source-id: 9702cc0bbfdb895ab4dcc059f4f57287950dc3f0