Add mechanism to conditionally disable property checking
Summary:
Sometimes we don't need to check certain properties.  For example, if a pass isn't in the pass list, then it may not establish properties which others conditionally depend on.

E.g. We don't care about source blocks in the outliner if InsertSourceBlocksPass was never in the pass list.  Similarly, we don't need to check the validity of class names vs. order if RenameClasses never ran.

This introduces a mechanism, via a PropertyName to function map, to conditionally enable or disable a checker for a property based on the results of that function. Initially it is a `bool(*)(ConfigFiles&)` function, but this could be changed in future.

Reviewed By: agampe

Differential Revision: D44966276

fbshipit-source-id: 44f174a41b98f53fa4e3f36c9463bcc7691a51cd