Fix array length population when analyzing FILLED_NEW_ARRAY
Summary:
The way we populate array length when analyzing a `OPCODE_FILLED_NEW_ARRAY` is incorrect. It now always set the array length to 0, which had no affect at the time. But as we made more use of the `ArrayNullnessDomain`, we can remove checks on arrays incorrectly.

Here we drop the array length population. As a result, the analysis does not construction a `ArrayNullnessDomain` anymore.

Reviewed By: wsanville, ssj933

Differential Revision: D45131427

fbshipit-source-id: 27b5feb97b457178d0544020345e74988c91362c