IntraDexCM does not preserve DexLimitsObeyed
Summary:
`IntraDexClassMergingPass` currently does not strictly preserve `DexLimitsObeyed` property. The reason being that when combining methods together, we do not perform the full inlining to remove the original methods from mergeable classes. The very next `IntraDexInliningPass` following closely will remove the leftover methods and re-establish the property.

The gap is when the property is not strictly preserved is only about 2 passes.

Reviewed By: beicy

Differential Revision: D45205835

fbshipit-source-id: ba730b180f1875e028712159fa3dc4ab35eade73