Resolve to external on simple path
Summary:
We have two paths to solve method/field refs in `ResolveRefsPass`. One is simply based on existing local ref. The other more complicated one (we added later on) utilizes `TypeInference`. We only resolve to external refs in the 2nd path. But doing that has no actual dependency on type inference.

In this change, we replicate the same resolve-to-external logic in the simple path without relying on the inferred type. Doing so actually increases the number of method/field references we can resolve to definition.

Reviewed By: NTillmann

Differential Revision: D45055770

fbshipit-source-id: a622cf42de5a48dd412c1b672ba6ab1de5e25c25