Add an --omit-sdk-tool-discovery flag to redex tools to skip SDK path based tool discovery
Summary:
As agampe pointed out, the utils.py script needs updating as well to stop attempting to locate tools via the sdk_path.

This change adds the --omit-sdk-tool-discovery flag to the redex tool to explicitly disable tool discovery via SDK path lookup. If set we'll respect any provided overrides and if not present attempt to source from the //third-party copies of the tool. If not set it'll use the existing SDK path lookup logic.

Reviewed By: agampe

Differential Revision: D45317398

fbshipit-source-id: 0544bb88d25a1a8cfdb27c8a2104dd70227bdc64