Make type inference based method refinement more lenient
Summary:
When resolving a method ref based on the local type inference info, we are being more conservative than we have to be. The existing logic bails for too many cases unnecessarily. For instance, we bail if the inferred type of the invocation target is an interface or if it `is_android_sdk_type`. They are not necessary anymore.

The existing logic can actually handle those skipped cases. By re-arranging the checks here, we can cover more method ref resolution cases. That in turn will leave less pure method ref untouched.

Reviewed By: agampe

Differential Revision: D45104713

fbshipit-source-id: 318e4a2e382586ff2e475d4f58d43a218528d2c9