Force inline virtual dispatch entries
Summary:
# Background

When merging classes together, we have to also merge methods together to preserve the semantics of the program. When doing that we don't always clean up all the leftovers. That can cause the method count to go up immediately after a Class Merging pass, even though overall the merging transformation reduces class/method/field counts.

The existing logic cleans up the merged methods at various places:
1. When merge ctors, we create a dispatch, force inline all entries in the dispatch, and we remove the original ctors after inlining.
2. When merging non-ctors and non-virtuals, we don't need to form a dispatch. We first dedup them and remove the ones replaced by the dedup process.
3. When merge true-virtuals, we form a dispatch. But right now we don't force inline each entries. That's where we leave a large number of methods to be clean up by the subsequent passes like Inlining or RMU.
4. Lastly, we purge merged classes with no dmethods left on the class.

# What's in this change

We need to clean up what's left by #3 above. Therefore, we force inline of the entries in a virtual dispatch we formed. After that, we only relocate the ones failed to be inlined to the merger class. Inlined entries remain on the merged class. They are then expected to be purged in #4 above.

Reviewed By: ssj933

Differential Revision: D45417489

fbshipit-source-id: 7574a91d560dd71f53b7db67a6eb79e972e8ee02