Expose unsafe usage types
Summary:
As title.

Move the internal enum in `EnumUpcastAnalysis` to `UnsafeType`. Add a callback for the event of rejections so that the caller knows the reasoning (instead of coupling with a large map).

Use the wired up callback to refactor & simplify the code, using less passthrough parameters for `EnumUpcastDetector` methods.

Reviewed By: NTillmann, thezhangwei

Differential Revision: D45632873

fbshipit-source-id: fd32632f335cd1601d90a4dbf941a45b07668c3d