introducing MethodSplittingPass
Summary:
The MethodSplittingPass is a new kind of outliner. While the existing InstructionSequenceOutlinerPass only targets relatively small instruction sequences at any point in the method, and the SplitHugeSwitchesPass can only outline certain code patterns in methods that start with a big switch, this new pass is designed to outline large portions of a method starting from any (well, there are some restrictions) block, or a set of switch cases associated with a switch. The main (current) limitation of the pass is that the outlined code spans from the selected block or switch cases to the exit blocks of the method. Or in other words, this pass outlines (almost) arbitrary suffices of methods. Multiple suffices can be outlined from the same method, and they can overlap, resulting in some code duplication in the outlined and primary methods. The pass selects suffices based on a cost function that tries to minimize such overlaps.

Benefits:
- better runtime performance: avoids hitting size limits that prevent on-device compilation
- even better runtime performance: splitting prefers hot-cold splits; this should help improve code locality
- reduced code size: allows inliner to inline more within given limits, by avoiding confusing the inliner's global cost function which isn't aware of the hard to predict limits

The implementation first computes a "reduced control-flow graph", where any strongly connected component is represented as a single reduced block. The reduced cfg is DAG, which makes the selection of code sections to outline more efficient.
We also split large blocks, to enable outlining some portion of them.
In the implementation, we call a splittable piece of code a "closure", in that it is the transitively reachable closure from some start block to the exit block of the reduced cfg.

(Some more relevant) current curable limitations, i.e. future opportunities:
- We currently don't support outlining from blocks nested in monitor instructions. (Seems rare.)
- We currently don't support outlining from blocks with throw-edges. (Seems rare.)
- We currently only consider the outermost strongly connected components. If we handle more diverse return types, we could run the algorithm for nested strongly connected components. (I have seen some remaining huge methods where we inlined a huge switch into a loop, so this might be worthwhile.)
- We impose various thresholds to balance the desire for small methods with the added overhead. While the particular thresholds seem reasonable and have already been tuned, this whole concept is a bit dubious. For example, instead of tracking what hit inlining thresholds and then chopping up those methods more, the inliner itself could use the new method-splitting functionality to chop up methods on demand, and then continue inlining.

Reviewed By: agampe

Differential Revision: D45342112

fbshipit-source-id: 7607a44059d26c2aff1c235a6a7bc80ffc539d70