switch inlining and related code to estimate_code_units
Summary:
The inliner has a built-in mechanism to prevent it from creating methods that are "too big", via soft and hard limits. The inliner used to not take into account switch or fill-array payloads, which effectively meant that it wasn't actually ensuring that the resulting method stayed under any limit!

This is getting fixed here. However, as I didn't increase the limits, this results in overall size regressions, as the inliner won't be able to inline as much as it used to.
(Don't worry, the next diffs up the stack improve the situation again.)

Reviewed By: agampe

Differential Revision: D44899340

fbshipit-source-id: f9f28834a93d3c907a0d9748dc3e8fe42fbd8529