Update debug-injector tool scripts to also take explicit path to the apksigner tool.
Summary:
Update debug-injector tool scripts to also take explicit path to the apksigner tool.

Tool seems to fail with the current configuration when run against an FB4A APK. We're conistently failing in the same way with this change though so it's getting the tool path correctly which is my main interest.

Reviewed By: agampe

Differential Revision: D45711787

fbshipit-source-id: 7193a682de20d46002e31df3c59d6b47159ff4ec