Revive the is_generated_code option for config based model
Summary:
For config based `ClassMergingPass`, the current behavior is we assume config based models are generated code. This is not always true. Having the merging target as generated code inadvertently relaxes mergeability checks specially for unsafe bytecode like `const-class`.

In this change, we reinstate this option to tighten the mergeability check when necessary.

Reviewed By: adicatana

Differential Revision: D46030978

fbshipit-source-id: 8e4a8f32b2272e74590bf953ec86036ec5d30b3e