Resolve Miranda method reference to interface method
Summary:
# Backgound

ResolveRefsPass currently does not handle `invoke-virtual` Miranda method refs. Here is an example:
```
interface I {
  I foo();
}
abstract class Base {
  I bar() {
    // ...
    return Base.foo();
  }
}

final class Sub {
  Override
  I foo() { ... }
}
```
There is a miranda ref in `Base.bar()`, which is `invoke-virtual Base.foo()`. Currently we do not resolve the pure method ref to a method def. If we update the proto of the methods in the virtual scope (e.g., rtype specialization), the pure ref will be left out.

# Solution

When resolving method refs, if the Resolver fails to identify the def with `MethodSearch::Virtual`, retry with `MethodSearch::InterfaceVirtual`. If the ref is resolved to an interface def, rewrite the `invoke-virtual` to `invoke-interface`.

Reviewed By: agampe

Differential Revision: D46168062

fbshipit-source-id: ec718b89032d04e8dd1f032e2f661a12895b9437