Option to gracefully handle duplicate cases in SwitchEquivFinder
Summary:
In order for D44735284 to work in practice without edge cases, it needs to be able to support messy realities, like redundant checks for the same case while making sure the one that would be encountered at runtime is chosen.

The existing DFS traversal that was already happening here is almost good enough to do it, but the iteration order needs to be fixed to process non-leafs before leafs.

Reviewed By: NTillmann

Differential Revision: D45893087

fbshipit-source-id: 0fe32e6d43fa8720382c612c7d2a9c93145e989b