Generalize SwitchEquivFinder for other types of const comparisons
Summary:
SwitchEquivFinder seems like a useful tool that could be used for a series of other types of equality checks too, not just int constants. For example, a series of equality checks against objects from `const-class` instructions (and maybe even some day string equality checks, who knows).

In particular, I like how this class tracks possible data originated in the non-leaf blocks that could be used downstream and organizes that in a useful way. Subsequent diffs in this stack will find that useful. So, this diff builds upon it to represent different types of keys, rather than just int.

Reviewed By: NTillmann

Differential Revision: D44735284

fbshipit-source-id: 2f63153d79c8075365766e33684750c13e487dea