Transform for methods with a lot of const-class equality checks
Summary: This is a new pass that attempts to find/transform methods that are doing a lot of equality checks against a given Java Class. DexTypes used in if/else blocks will get ordered, given an ordinal, then munged into a data structure. A method that conforms to expectations will then get transformed to consult the data structure to get the ordinal, and then switch on it to go to the original dest blocks. This is meant to achieve the same effect as the original code without being a class loading bomb.

Reviewed By: NTillmann

Differential Revision: D44773185

fbshipit-source-id: d87be5159ab947293811d740ce96a14be8abb8a5