Resolve array clone reference in ResolveRefsPass
Summary:
The existing logic that rewrites a method ref to array clone of a reference type resides in `ReBindRefsPass`.
It rebinds a method ref from `[LFoo;.clone:()Ljava/lang/Object;` to `[Ljava/lang/Object;.clone:()Ljava/lang/Object;`. Neither `clone` method is a method def, but rewriting to the one on `[Ljava/lang/Object;` is the best we can do in terms of minimizing unique method refs.

However, `ReBindRefsPass` is rather late in the pipeline. Having this resolved early on helps to minimize the resolvable pure method refs. This change relocates this handling logic to `ResolveRefsPass` so that we have less resolvable pure refs to worry about in subsequent passes.

Reviewed By: NTillmann

Differential Revision: D46238541

fbshipit-source-id: 712d5266afc483e5bfd2c9b8ee00aee618b0e65f