use is_cfg_legacy() to replace is_editable_cfg_friendly()
Summary:
In the past, we set is_editable_cfg_friendly() to True to mark a pass is fully updated to editable cfg.   Then for newly added pass, people need to explicitly to override the method, and later, once all passes are updated to editable cfg, we need to delete all is_editable_cfg_friendly() as well.
Therefore, we would like to use an opposite way : for passes we haven't explicitly updated to editable cfg, we set is_cfg_legacy() to be true.  Once a pass is checked and updated to editable cfg, the override method is_cfg_legacy() is just deleted.  Also, for newly added passes, as long as it uses editable cfg,  there is no need to override this method.
This change is behavior preserving.
Note: for passes which are marked is_cfg_legacy() to be true, just mean that, we haven't double check it cfg usage yet.

Reviewed By: NTillmann

Differential Revision: D46325638

fbshipit-source-id: 9b9b0e96843955cf9ef1f016a1cd91282843b679