Be more conservative in RemoveRedundantCheckCasts
Summary:
The optimization only checked the class hierarchy that is loaded. This includes Android SDK classes, at the level of the SDK.

Code that is MinSDK sensitive may rely on `check-cast` instructions to be type-correct and pass the Android verifier. Ensure that if a class is not in the MinSDK, or a class chain is different, that a check-cast is not elided because it would always succeed on new enough API levels.

Reviewed By: NTillmann

Differential Revision: D46923924

fbshipit-source-id: e1831554256a9c2e4aca499d5fd1401490e29d41